Update for TS7
I am testing Typescript 7's JS support, which I've largely rewritten
during the switch to Go. The new code processes tags by converting them
into synthetic nodes just after parsing their host node. For unified,
that means that `@template` tags can't be shared between multiple
`@overload` tags like before.

This PR also has to convert function declarations with `@type` on them
to variables initialised with a function expression. TS7 doesn't
currently support this feature on function declarations--but both unified and
svelte use this feature a lot, so I think we'll need to add it before
TS7's JS features are ready. At that point the shuffling of functions in
the test files won't be needed anymore.

In addition, https://github.com/microsoft/typescript-go/pull/1444 makes
it possible to have rest parameters in `@overload`, and is needed for
this change to work in TS7, so it won't compile with the nightly until
that PR is merged.

Because TS7 is quite a way off, I don't know whether you'll want to take this PR. I created it to see how hard it would be to update popular JS code that uses TS for checking.